remove macro support and transfer everything to API
